finish the proof for suspend.
There is something I am still missing about how they manipulate the
logical state of the cells. In the suspend function I want to change the
global list state when reading the pointer. For that I need to use a
big_sepL_lookup_alter lemma which is never used otherwise but I am not
sure how they are updating the list otherwise.